feat: add bootOrder to vm block device refs status

+++ images/virtualization-artifact/pkg/controller/kvbuilder/kvvm_utils.go
@@ -194,7 +194,7 @@ func applyBlockDeviceRefs(
 		var kvBootOrder uint
 		if hasExplicitBootOrder {
 			if bd.BootOrder != nil {
-				kvBootOrder = uint(*bd.BootOrder)
+				kvBootOrder = *bd.BootOrder
 			}
 		} else {
 			kvBootOrder = uint(i) + 1
